requirements.txt
google-generativeai>=0.5.0
opencv-python>=4.10.0.0
python-dotenv>=1.0.1
pillow>=10.4.0
numpy>=1.26.0
